+++ rules/standard/signbeaconattestations.go
@@ -17,6 +17,7 @@ import (
 	"bytes"
 	"context"
 	"errors"
+	"fmt"
 	"time"
 
 	"github.com/attestantio/dirk/rules"
@@ -117,12 +118,12 @@ func (s *Service) fetchSignBeaconAttestationStates(ctx context.Context, pubKeys 
 	return states, nil
 }
 
-func (s *Service) runSignBeaconAttestationChecks(ctx context.Context, metadata *rules.ReqMetadata, req *rules.SignBeaconAttestationData, state *signBeaconAttestationState) rules.Result {
+func (s *Service) runSignBeaconAttestationChecks(_ context.Context, metadata *rules.ReqMetadata, req *rules.SignBeaconAttestationData, state *signBeaconAttestationState) rules.Result {
 	log := log.With().Str("client", metadata.Client).Str("account", metadata.Account).Str("rule", "sign beacon attestation").Logger()
 
 	// The request must have the appropriate domain.
 	if !bytes.Equal(req.Domain[0:4], e2types.DomainBeaconAttester[:]) {
-		log.Warn().Msg("Not approving non-beacon attestation due to incorrect domain")
+		log.Warn().Str("domain", fmt.Sprintf("%#x", req.Domain)).Msg("Not approving non-beacon attestation due to incorrect domain")
 		return rules.DENIED
 	}