+++ services/signer/standard/signbeaconproposal.go
@@ -1,4 +1,4 @@
-// Copyright © 2020 Attestant Limited.
+// Copyright © 2020, 2022 Attestant Limited.
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -23,6 +23,8 @@ import (
 	"github.com/attestantio/dirk/services/checker"
 	"github.com/attestantio/dirk/services/ruler"
 	spec "github.com/attestantio/go-eth2-client/spec/phase0"
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/attribute"
 )
 
 // SignBeaconProposal signs a proposal for a beacon block.
@@ -36,12 +38,15 @@ func (s *Service) SignBeaconProposal(
 	core.Result,
 	[]byte,
 ) {
+	ctx, span := otel.Tracer("attestantio.dirk.services.signer.standard").Start(ctx, "SignBeaconProposal")
+	defer span.End()
 	started := time.Now()
 
 	if credentials == nil {
 		log.Warn().Msg("No credentials supplied")
 		return core.ResultFailed, nil
 	}
+	span.SetAttributes(attribute.String("client", credentials.Client))
 
 	log := log.With().
 		Str("request_id", credentials.RequestID).
@@ -77,12 +82,29 @@ func (s *Service) SignBeaconProposal(
 		return core.ResultDenied, nil
 	}
 
+	if e := log.Trace(); e.Enabled() {
+		e.Str("domain", fmt.Sprintf("%#x", data.Domain)).
+			Uint64("slot", data.Slot).
+			Uint64("proposer_index", data.ProposerIndex).
+			Str("parent_root", fmt.Sprintf("%#x", data.ParentRoot)).
+			Str("state_root", fmt.Sprintf("%#x", data.StateRoot)).
+			Str("body_root", fmt.Sprintf("%#x", data.BodyRoot))
+		if len(accountName) > 0 {
+			e.Str("account", accountName)
+		}
+		if len(pubKey) > 0 {
+			e.Str("pubkey", fmt.Sprintf("%#x", pubKey))
+		}
+		e.Msg("Data to sign")
+	}
+
 	wallet, account, checkRes := s.preCheck(ctx, credentials, accountName, pubKey, ruler.ActionSignBeaconProposal)
 	if checkRes != core.ResultSucceeded {
 		s.monitor.SignCompleted(started, "proposal", checkRes)
 		return checkRes, nil
 	}
 	accountName = fmt.Sprintf("%s/%s", wallet.Name(), account.Name())
+	span.SetAttributes(attribute.String("account", accountName))
 	log = log.With().Str("account", accountName).Logger()
 
 	// Confirm approval via rules.
@@ -104,6 +126,12 @@ func (s *Service) SignBeaconProposal(
 		log.Error().Str("result", "failed").Msg("Rules check failed")
 		s.monitor.SignCompleted(started, "proposal", core.ResultFailed)
 		return core.ResultFailed, nil
+	case rules.UNKNOWN:
+		log.Error().Str("result", "failed").Msg("Rules check indeterminate result")
+		s.monitor.SignCompleted(started, "proposal", core.ResultFailed)
+		return core.ResultFailed, nil
+	case rules.APPROVED:
+		//  Nothing to do.
 	}
 
 	// Create a spec version of the beacon block header to obtain its hash tree root.