+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Development
+  - add Amazon Secrets Manager (ASM) confidant
+  - print command errors to stderr
+  - provide clearer error message on configuration issues
+  - use bytes rather than string on slashing protection import
+  - provide full information on the data being signed with trace logging
+  - update tracing implementation to use opentelemetry
+
 # Version 1.1.0
   - garbage collect the slashing database on startup to reduce on-disk size
   - provide release metric in `dirk_release`